REST API endpoints - 1: CommonController - /userprofile/{userId}
Features implemented:
User profile does not exist in the database: throw 'UserNotFoundException' with the message code -'USR-001' and message -'User with entered uuid does not exist'.
User profile exist in the database: Return all the details of the user from the database in the JSON response with the corresponding HTTP status.