REST API endpoints - 1: QuestionController - /question/create
Features implemented:
Access token of the user does not exist in the database: throw 'AuthorizationFailedException' with the message code - 'ATHR-001' and message - 'User has not signed in'.

Access token of the user does not exist in the database: Throw exception - 'ATHR-001' and message - 'User has not signed in'.
User has signed out - Throw exception - 'ATHR-002' and message -'User is signed out.Sign in first to post a question'.
Else, save the question in the database and return the 'uuid' of the question and message 'QUESTION CREATED' in the JSON response with the corresponding HTTP status.